Add derived client
The derived client works as expected but I still need to work on the
worker implementation side. I want to create a function that creates
the worker and gets the worker context as an argument, similar to how
the constructor for observable works.

Something like:

```
function createWorker<T extends WorkerContract<T>>(
  (clientId:string, subscribe: WorkerSubscribe)=> T
)
```